First, created the start of the navigation bar. It still needs to be finished as it is missing the links to the other pages. This hasn't been implemented because there are currently no other pages to implement.
For Style.css I have implemented some more classes to help style the navbar along with the whole website. This means the navbar with its content sits in the top-middle, and I have added all the colours needed for the website.

Finally, for Reset.css I have changed the defatul font to something that looks nicer and more clean.